can edit card audio, and old audio will be gracefully deleted when card is updated

+++ lib/classes/card_audio.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:K9_Karaoke/services/rest_api.dart';
+import 'package:uuid/uuid.dart';
 
 class CardAudio {
   String fileId;
@@ -9,14 +10,20 @@ class CardAudio {
   List amplitudes;
 
   CardAudio({
-    this.fileId,
     this.filePath,
     this.bucketFp,
     this.amplitudes,
-  });
+  }) {
+    this.fileId = Uuid().v4();
+  }
+
+  void deleteFile() {
+    if (File(filePath).existsSync()) File(filePath).deleteSync();
+  }
 
   Future<void> delete() async {
-    await RestAPI.deleteCardAudio(fileId);
-    if (File(filePath).existsSync()) File(filePath).deleteSync();
+    print("Deleting old card audio");
+    if (bucketFp != null) await RestAPI.deleteCardAudio(fileId);
+    deleteFile();
   }
 }
